Remove encoder_padding_mask
The encoder_padding_mask is not used at all in the RetNet architecture.
As for the Transformer architecture, the encoder_padding_mask is only
used when an encoder_out is given, which we do not give. This also makes
logical sense not to use this mask because of our decoder-only stack.