Slight Tweaks + Remove Support For Non-BPE Tokenizers
Remove support for non-BPE tokenizers. Could be implemented at another
time if desired. Also remove the end-of-sequence tokens due to
problems with data processing and this token not being specified in
the original RetNet paper. Fix bug where tokenizer was only training
on the word "text" instead of the actual dataset.