Update Struct __setattr__ to update YAML
When setting attributes of config with changes like `config.some_attr`, the new attribute would appear in it's properties (`dir(config)`) but not in its config_dict property, like it should. The config dict is used for printing out parameters and naming files, so it's important to have it updating alongside the config properties.